Handle optional fields in profile
