+++ src/components/extensions/membership/MembershipProfile.vue
@@ -1,16 +1,8 @@
 <script setup lang="ts">
 import { useMembershipStore } from "@/stores/membership";
 import { useMenuStore } from "@/stores/menu";
-import { useUserStore } from "@/stores/user";
 import { storeToRefs } from "pinia";
-import { useI18n } from "vue-i18n";
-import InputText from "primevue/inputtext";
-import { ref, watch } from "vue";
-import { useVuelidate } from "@vuelidate/core";
-import { required } from "@vuelidate/validators";
-import { useToast } from "primevue/usetoast";
-import Toast from "primevue/toast";
-import RadioButton from "primevue/radiobutton";
+import MembershipProfileDetail from "./MembershipProfileDetail.vue";
 
 const menuStore = useMenuStore();
 menuStore.setTitle("Membership");
@@ -18,272 +10,35 @@ menuStore.setTitle("Membership");
 const membershipStore = useMembershipStore();
 membershipStore.getMembershipSchema();
 membershipStore.getMembership();
-
-const userStore = useUserStore();
-const { user } = storeToRefs(userStore);
-
-const { t } = useI18n();
-const selectedGender = ref();
-const toast = useToast();
-const submitted = ref(false);
 const { membership, membershipSchema } = storeToRefs(membershipStore);
 
-watch(membership, (value) => {
-  selectedGender.value = value ? value["gender"] : "";
-});
-
-async function openKeycloakAccount() {
-  const accountUrl = await user.value?.accountUrl;
-  window.open(accountUrl, "_blank");
-}
-
-// Vuelidate block
-// Required fields need to be added manually and will be checked afterwards, if each field is contained in membershipSchema
-const rules: any = {}
-for (var key in membershipSchema.value) {
-  if (membershipSchema.value[key].required) {
-    if (membershipSchema.value[key].condition?
-    rules[key] = { required }
-  }
-}
-
-// const rules = {
-//   first_name: { required },
-//   last_name: { required },
-//   birth_date: { required
-
-//   gender: { required },
-//   occupation: { required },
-//   address_street: { required },
-//   address_number: { required },
-//   address_postcode: { required },
-//   address_city: { required },
-//   address_country: { required },
-// };
-const v$ = useVuelidate(rules, membership);
-watch(membershipSchema, () => {
-  const dict: Array<Array<string | any>> = Object.entries(
-    JSON.parse(JSON.stringify(membershipSchema.value))
-  );
-  for (let i = 0; i < dict.length; i++) {
-    if (dict[i][1].required) {
-      if (!Object.keys(rules).includes(dict[i][0])) {
-        throw TypeError("Missing required field in rules");
-      }
-    }
-  }
-});
-
-function updateInputText(event: any, key: string | number) {
-  if (membership.value) {
-    membership.value[key as keyof typeof membership.value] = event.target.value;
-  }
-}
-
-function updateRadioButton(label: any, key: string) {
-  if (membership.value) {
-    membership.value[key as keyof typeof membership.value] = label;
-  }
-}
-
-function capitalized(name: string) {
-  const capitalizedFirst = name[0].toUpperCase();
-  const rest = name.slice(1);
-  return capitalizedFirst + rest;
-}
-
-function convertDate(value: string) {
-  //convert date to german format from https://stackoverflow.com/a/2086843/19932351
-  let datePart = value.match(/\d+/g);
-  if (datePart) {
-    let year = datePart[0].substring(2);
-    let month = datePart[1];
-    let day = datePart[2];
-    return day + "." + month + "." + year;
-  }
-}
-
-function returnErrorMessage(key: any) {
-  if (Object.keys(rules).includes(key)) {
-    return v$.value[key].required.$message;
-  } else {
-    return false;
-  }
-}
-
-function isInvalid(key: any) {
-  if (Object.keys(rules).includes(key)) {
-    return (
-      (v$.value[key].$invalid && submitted) || v$.value[key].$pending.$response
-    );
-  } else {
-    return false;
-  }
-}
-
-async function save() {
-  submitted.value = true;
-  const isFormCorrect = await v$.value.$validate();
-  if (!isFormCorrect) {
-    toast.add({
-      severity: "error",
-      summary: t("Error"),
-      detail:
-        typeof v$.value.$errors[0].$message === "string"
-          ? t(v$.value.$errors[0].$message)
-          : "",
-      life: 5000,
-    });
-    return;
-  } else if (membership.value) {
-    try {
-      await membershipStore.updateMembership(membership.value);
-      toast.add({
-        severity: "success",
-        summary: t("Profile updated"),
-        life: 5000,
-      });
-    } catch (e: any) {
-      toast.add({
-        severity: "error",
-        summary: t("Update failed"),
-        detail: `Request-id: "${e?.response?.headers["x-request-id"]}" `,
-      });
-    }
-  } else {
-    throw new TypeError("Membership value is empty");
-  }
-}
-
-function schemaToPrime(choices: any) {
-  let schema: Array<Object> = [];
-  Object.keys(choices).forEach((key) => {
-    schema.push({ name: choices[key], code: choices[key] });
-  });
-  return schema;
-}
 </script>
 
 <template>
-  <div>
-    <Toast />
-    <div
-      v-if="membershipSchema === null && membership === null"
-      class="loading"
-    >
-      <h2>{{ $t("Loading members") }}</h2>
+  <div class="members-wrapper">
+    <div v-if="membership === null || membershipSchema === null" class="loading">
+      <h2>{{ $t("Loading membership data") }}</h2>
     </div>
     <div v-else class="members-table">
-      <div
-        v-for="(value, key) in membershipSchema"
-        :key="value ? key + value.input_type : key"
-        class="field"
-      >
-
-        <!-- Read only fields -->
-        <div v-if="value.read_only">
-          <h3>{{ t(value?.label) }}</h3>
-          <div v-if="value.input_type === 'date'">
-            <InputText
-              disabled
-              :value="membership ? membership[key as keyof typeof membership] ? convertDate(membership[key as keyof typeof membership] as string) : '' : ''"
-            />
-          </div>
-          <div v-else-if="value?.input_type === 'email'">
-            <InputText
-              disabled
-              :value="membership ? membership[key as keyof typeof membership] : ''"
-            />
-            <div class="changeMailButton">
-              <ButtonPrime
-                :label="t('Change mail or password')"
-                icon="pi pi-envelope"
-                @click="openKeycloakAccount()"
-                autofocus
-              />
-            </div>
-          </div>
-          <div v-else>
-            <InputText
-              disabled
-              :value="membership ? membership[key as keyof typeof membership] : ''"
-            />
-          </div>
-        </div>
-
-        <!-- Editable fields -->
-        <div v-else>
-          <div v-if="value?.input_type === 'text'">
-            <h3 v-if="value.required">{{ t(value?.label) }}*</h3>
-            <h3 v-else>{{ t(value?.label) }}</h3>
-
-            <InputText
-              id="user-attr-{{value}}"
-              :type="value?.input_type"
-              aria-describedby="user-attr-{{value}}-help"
-              @change="updateInputText($event, key)"
-              :value="membership ? membership[key as keyof typeof membership] : ''"
-              :class="{ 'p-invalid': isInvalid(key) }"
-            />
-            <br />
-            <span v-if="isInvalid(key)" class="p-error">{{
-              returnErrorMessage(key)
-            }}</span>
-          </div>
-          <div v-else-if="value?.input_type === 'radio'">
-            <div v-if="key === 'gender'" class="w-56">
-              <h3 v-if="value.required">{{ t(value?.label) }}*</h3>
-              <h3 v-else>{{ t(value?.label) }}</h3>
-              <br />
-              <div
-                v-for="category of (schemaToPrime(value.choices) as any)"
-                class="field-radiobutton"
-                :key="category.name"
-              >
-                <RadioButton
-                  :inputId="category.name"
-                  name="key"
-                  :value="category.name"
-                  v-model="selectedGender"
-                  @click="updateRadioButton(category.name, key)"
-                />
-                <label :for="category.name" class="genderLabel">{{
-                  t(capitalized(category.name))
-                }}</label>
-              </div>
-              <span v-if="isInvalid(key)" class="p-error">{{
-                returnErrorMessage(key)
-              }}</span>
-            </div>
-          </div>
-          <div v-else>
-            <h3>{{ value?.label }} Else-Case!!</h3>
-            <InputText
-              id="user-attr-{{value}}"
-              :type="value?.input_type"
-              aria-describedby="user-attr-{{value}}-help"
-              :value="value"
-              :disabled="typeof key === 'string' ? key === 'id' : false"
-            />
-          </div>
-        </div>
-        <br />
-      </div>
-      <ButtonPrime
-        :label="t('Save')"
-        icon="pi pi-check"
-        @click="save()"
-        autofocus
-      />
+      <MembershipProfileDetail :membership="membership" :membershipSchema="membershipSchema" />
     </div>
   </div>
 </template>
 
-<style>
-.genderLabel {
-  padding-left: 10px;
+<style scoped>
+.error-message {
+  border: 1px solid red;
+  padding: 10px;
+  border-radius: 5px;
 }
-.changeMailButton {
-  padding-top: 10px;
+
+h1 {
+  font-weight: 500;
+  font-size: 2.6rem;
+  top: -10px;
+}
+
+h3 {
+  font-size: 1.2rem;
 }
 </style>
